WaypointNav class created
