+++ src/navigation.cpp
@@ -17,3 +17,13 @@ float headingToWaypoint(Pose target, Pose current){
 
     return relativeHeading;
 }
+
+float distanceToWaypoint(Pose target, Pose current){
+    //returns distance 'as the crow flies' to the target pose 
+    float distance;
+    //hypotenuse of dx, dy triangle gives distance, using h^2=x^2+y^2
+    distance = sqrt(powf((target.x-current.x),2) + powf((target.y-current.y),2));
+//    display.println(" ");
+//    display.printf("distance: %2.2f", distance);
+    return distance;
+}
