+++ src/utils.cpp
@@ -22,14 +22,3 @@ float wrapTwoPi(float angle) {
     while (angle < -M_PI) angle += TWO_PI;
     return angle;
 }
-
-float PID::update(float targetValue, float currentValue, float previousValue, float timeInterval){
-    // uses proportional, differential and integral gains
-    // to calcualte the response for a given error from target
-    // values floats
-    float error = currentValue - targetValue;
-    float change = currentValue - previousValue;
-    cumulativeError += error;
-    float output = pGain * error - dGain * change / timeInterval + iGain * cumulativeError;
-    return output;
-};
